Fixed the tests, moving them to the appropriate area.
